Fix class naming
